microservice/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.7
django-extensions==3.2.0
django-taggit==3.0.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
urllib3==2.0.3
gunicorn
psycopg2-binary
